FEAT : domain 디렉토리 구조 적용 (#8)
