--- src/main/java/com/example/betteriter/config/MainConfig.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.config;
-
-public class MainConfig {
-}
