--- src/main/java/com/example/betteriter/controller/MainController.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.controller;
-
-public class MainController {
-}
