--- src/main/java/com/example/betteriter/domain/entity/MainEntity.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.domain.entity;
-
-public class MainEntity {
-}
