--- src/main/java/com/example/betteriter/domain/repository/MainRepository.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.domain.repository;
-
-public interface MainRepository {
-}
