--- src/main/java/com/example/betteriter/exception/MainException.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.exception;
-
-public class MainException {
-}
