+++ src/main/java/com/example/betteriter/global/common/entity/BaseEntity.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.global.common.entity;
+
+public class BaseEntity {
+}
