+++ src/main/java/com/example/betteriter/global/common/response/BaseResponse.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.global.common.response;
+
+public class BaseResponse {
+}
