+++ src/main/java/com/example/betteriter/global/config/MainConfig.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.global.config;
+
+public class MainConfig {
+}
