+++ src/main/java/com/example/betteriter/global/error/ErrorResponse.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.global.error;
+
+public class ErrorResponse {
+}
