+++ src/main/java/com/example/betteriter/global/error/GloblaExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.global.error;
+
+public class GloblaExceptionHandler {
+}
