+++ src/main/java/com/example/betteriter/infra/Email.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.infra;
+
+public class Email {
+}
