+++ src/main/java/com/example/betteriter/main/controller/MainController.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.controller;
+
+public class MainController {
+}
