+++ src/main/java/com/example/betteriter/main/dao/MainDao.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.dao;
+
+public class MainDao {
+}
