+++ src/main/java/com/example/betteriter/main/domain/entity/MainEntity.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.domain.entity;
+
+public class MainEntity {
+}
