+++ src/main/java/com/example/betteriter/main/domain/repository/MainRepository.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.domain.repository;
+
+public interface MainRepository {
+}
