+++ src/main/java/com/example/betteriter/main/dto/MainDto.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.dto;
+
+public class MainDto {
+}
