+++ src/main/java/com/example/betteriter/main/exception/MainException.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.exception;
+
+public class MainException {
+}
