+++ src/main/java/com/example/betteriter/main/service/MainService.java
@@ -1,0 +1,4 @@
+package com.example.betteriter.main.service;
+
+public class MainService {
+}
