--- src/main/java/com/example/betteriter/service/MainService.java
@@ -1,4 +1,0 @@
-package com.example.betteriter.service;
-
-public class MainService {
-}
